+++ enum_ext/example/main.dart
@@ -2,7 +2,9 @@ import 'package:enum_ext/enum_ext.dart';
 
 part 'main.g.dart';
 
-@EnumExt()
+@EnumExt(
+  conditionalGetters: false,
+)
 enum HttpResponse {
   @EnumExtValue(200)
   ok,