+++ enum_ext/lib/src/enum_annotations.dart
@@ -1,9 +1,24 @@
 class EnumExt {
+  /// Whether to create conditional value getters or not
+  ///
+  ///```dart
+  /// @EnumExt()
+  /// enum Gender {
+  ///   male,
+  ///   female,
+  ///   other,
+  /// }
+  /// ```
+  ///
+  /// Now you can use value checking getters like `gender.isMale`, `gender.isNotMale` etc. instead of `gender == Gender.male` or `gender != Gender.male`
+  ///
+  final bool conditionalGetters;
+
   /// __EnumExt__ annotation makes `enum_ext_gen` package to generate some useful methods
   /// in order to rescue us from __switch-case__ and __if-else__ blocks.
   ///
   ///
-  const EnumExt();
+  const EnumExt({this.conditionalGetters = true});
 }
 
 class EnumExtValue {