+++ example/bin/enums.g.dart
@@ -45,7 +45,9 @@ extension HttpResponseExt on HttpResponse {
       case HttpResponse.notFound:
         return notFound == null ? orElse(this) : notFound(this);
       case HttpResponse.internalServerError:
-        return internalServerError == null ? orElse(this) : internalServerError(this);
+        return internalServerError == null
+            ? orElse(this)
+            : internalServerError(this);
     }
   }
 
@@ -106,7 +108,9 @@ extension HttpResponseExt on HttpResponse {
       case HttpResponse.notFound:
         return notFound == null ? orElse(this) : notFound(this);
       case HttpResponse.internalServerError:
-        return internalServerError == null ? orElse(this) : internalServerError(this);
+        return internalServerError == null
+            ? orElse(this)
+            : internalServerError(this);
     }
   }
 }