+++ enum_ext_gen/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.2.0
+
+- Generating value checking getters like `is...` and `isNot...`
 ## 1.1.0
 
 - Added example